requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.6.1
elevenlabs==0.2.26
python-multipart==0.0.6
websockets==12.0
asyncio-mqtt==0.13.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
opencv-python==4.8.1.78
mediapipe==0.10.7
langchain==0.0.350
langchain-openai==0.0.2
redis==5.0.1
celery==5.3.4 